+++ src/main/java/com/javarush/jira/bugtracking/DashboardUIController.java
@@ -1,6 +1,5 @@
 package com.javarush.jira.bugtracking;
 
-import com.javarush.jira.bugtracking.task.TaskService;
 import com.javarush.jira.bugtracking.to.SprintTo;
 import com.javarush.jira.bugtracking.to.TaskTo;
 import lombok.AllArgsConstructor;
