+++ src/main/java/com/javarush/jira/bugtracking/internal/repository/UserBelongRepository.java
@@ -1,6 +1,7 @@
 package com.javarush.jira.bugtracking.internal.repository;
 
 import com.javarush.jira.bugtracking.internal.model.UserBelong;
+import com.javarush.jira.bugtracking.to.ObjectType;
 import com.javarush.jira.common.BaseRepository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -8,5 +9,6 @@ import java.util.List;
 
 @Transactional(readOnly = true)
 public interface UserBelongRepository extends BaseRepository<UserBelong> {
-    List<UserBelong> getUserBelongsByObjectIdAndUserTypeCode(Long objectId, String userTypeCode);
+    List<UserBelong> getUserBelongsByObjectIdAndUserTypeCodeAndObjectType(
+            Long objectId, String userTypeCode, ObjectType objectType);
 }
